remove `do`-notation from `let .. in` uses
Not needed, and probably better to defer introduction of it until later.

It's still present in the actual `runMain` flow, but I'll hope that
people are more concerned with the `let .. in` at this juncture to
notice or care. We already have to handwave about `<-` and so on at that
point so we can handwave about or ignore `do` for now.