rename `Holding` constructor to `Holder`
Doesn't serve pedagogical purpose considering the road to this point in
the material; the students are already painfully aware that constructors
can be named differently than their types at this point. Better to
illustrate that when you have one constructor, the default is to name it
the same as the type.